+++ src/components/MainContainer/MainContainer.scss
@@ -84,6 +84,14 @@ input::placeholder {
     overflow: hidden !important;
 }
 
+.app-version {
+    position: absolute;
+    top: 20px;
+    left: 25px;
+    font-size: 10px;
+    color: $grey4;
+}
+
 @media screen and (max-width: 1700px) {
     .main-container .right-panel {
         padding: 0 100px;
