+++ src/components/Record/Record.tsx
@@ -5,7 +5,7 @@ import CheckmarkButton from '../Buttons/CheckmarkButton'
 import { MdExpandLess, MdExpandMore } from 'react-icons/md'
 import RecordMenu from '../RecordMenu/RecordMenu'
 import Editable from './Editable'
-import { updateTask } from '../../utils/taskService'
+import { updateTask } from '../../services/taskService'
 import RecordList from '../RecordList/RecordList'
 import { store, useEvent, notify, Events } from '../../classes/Store'
 import metadata from '../../classes/Metadata'
