--- src/components/Sidebar/Sidebar.scss
@@ -1,14 +1,0 @@
-@import '../../styles/variables';
-
-// .left-panel {
-//     position: relative;
-//     min-width: 400px;
-//     padding: 0 6px;
-//     border-right: 1px solid $grey3; 
-//     background-color: $grey1;
-//     overflow: auto;
-
-//     .close-menu-btn {
-//         display: none;
-//     }
-// }
