+++ src/version.js
@@ -1,0 +1,1 @@
+export const APP_VERSION = '0.2.0'
